Basic authentication support
When prometheus is started it can be passed arguments to facilitate
basic authentication.  These include:

        --web.auth-enabled
        --web.basicauth.username
        --web.basicauth.hash
        --web.basicauth.salt
        --web.basicauth.iterations

When a REST request comes it the username and password will be in the
header.  The username must match the one specified to
web.basicauth.username.  The password will be hashed using the salt and
the number of iterations using pbkdf2 and the result must match
web.basicauth.hash.