Basic authentication support
When prometheus is started it can be passed the name of a file
containing information used to do basic authentication of REST requests.

        --web.basicauth.config <file>

The <file> contains the expected username and hash values as well as the
salt and itertions needed to compute the hash.
A REST request is authenticated when the username in the request matches
the expected username and the pbkdf2 generated value matches the
expected has.